task-update

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // because When user enters 1000.25 in the form field and saves - it's saved as '1000.25',But if user enters 10000.25 - it's saved as '10000.2'. So it cuts the last digit if the number is over 9999. the solution of  precise representation and accuracy, especially for financial data, decimal datatype.
+            $table->decimal('commission_rate');
             $table->string('discount_code');
             $table->timestamps();
         });
